adding ConstructorFailed exception testing, refactoring test class to load data providers into single location
