+++ composer.json
@@ -31,5 +31,10 @@
         "psr-4": {
             "ParagonIE\\EasyDB\\": "src"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "ParagonIE\\EasyDB\\Tests\\": "tests"
+        }
     }
 }
